refactor(velocity_smoother): rename to include/autoware/{package_name}
