+++ planning/behavior_velocity_planner/autoware_behavior_velocity_planner_common/include/autoware/behavior_velocity_planner_common/planner_data.hpp
@@ -18,8 +18,8 @@
 #include "autoware_route_handler/route_handler.hpp"
 
 #include <autoware/behavior_velocity_planner_common/utilization/util.hpp>
+#include <autoware/velocity_smoother/smoother/smoother_base.hpp>
 #include <autoware_vehicle_info_utils/vehicle_info_utils.hpp>
-#include <autoware_velocity_smoother/smoother/smoother_base.hpp>
 
 #include <autoware_map_msgs/msg/lanelet_map_bin.hpp>
 #include <autoware_perception_msgs/msg/predicted_objects.hpp>
