+++ planning/behavior_velocity_planner/autoware_behavior_velocity_planner_common/src/utilization/trajectory_utils.cpp
@@ -16,7 +16,7 @@
 #include "motion_utils/trajectory/conversion.hpp"
 
 #include <autoware/behavior_velocity_planner_common/utilization/trajectory_utils.hpp>
-#include <autoware_velocity_smoother/trajectory_utils.hpp>
+#include <autoware/velocity_smoother/trajectory_utils.hpp>
 #include <motion_utils/trajectory/trajectory.hpp>
 #include <rclcpp/rclcpp.hpp>
